refactor trace collector tests
